wechat-monitor/requirements.txt
# RSS解析
feedparser==6.0.11

# HTTP请求
requests==2.31.0

# HTML处理
beautifulsoup4==4.12.3
lxml==5.2.1
html2text==2024.2.26

# AI SDK
anthropic==0.25.1

# 配置文件
pyyaml==6.0.1

# 日志
python-dateutil==2.9.0

# CSV处理
pandas==2.2.2

# 表格显示（用于查询工具）
tabulate==0.9.0
